requirements.txt
Pillow
plexapi
